[Clang][Sema] Moved test to appropriate file
